requirements.txt
fastapi[all]
uvicorn[standard]
requests==2.28.2
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
python-dotenv==1.0.0